% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{correctedRegression}
\alias{correctedRegression}
\title{Compute the moment corrected estimator for the simple linear regression.}
\usage{
correctedRegression(D, W, sigmaU)
}
\arguments{
\item{D}{the the response vector}

\item{W}{the the predictor vector}

\item{sigmaU}{the measurement uncertainty vector for the predictor}
}
\value{
the estamated regression coefficients
}
\description{
Compute the moment corrected estimator for the simple linear regression.
}
\examples{
library(measError)
set.seed(100)
n = 400
beta0 = 1
beta1 = 2
sigmaU = runif(n, 0.4, 0.6)
sigmaQ = runif(n, 0.1, 0.2)
sigmaE = 0.2
simuData = getData(n, beta0, beta1, sigmaU, sigmaQ, sigmaE)
(betaHat1 = naiveRegression(simuData$D, simuData$W))
(betaHat2 = correctedRegression(simuData$D, simuData$W, sigmaU))
}
