% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData.R
\name{getData}
\alias{getData}
\title{get simulation data}
\usage{
getData(n, beta0, beta1, sigmaU, sigmaQ, sigmaE)
}
\arguments{
\item{n}{the sample size}

\item{beta0}{the model intercept}

\item{beta1}{the model slope}

\item{sigmaU}{the measurement uncerntainty for X}

\item{sigmaQ}{the measurement uncertainty for Y}

\item{sigmaE}{the model error}
}
\value{
the generated data as a data frame
}
\description{
get simulation data
}
