% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{naiveRegression}
\alias{naiveRegression}
\title{Compute the ordinary least squares.}
\usage{
naiveRegression(D, W)
}
\arguments{
\item{D}{the the response vector}

\item{W}{the the predictor vector}
}
\value{
the estamated regression coefficients
}
\description{
Compute the ordinary least squares.
}
